+++ src/main/java/com/ecommerce/project/payload/SellerOrderDTO.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -16,6 +17,6 @@ public class SellerOrderDTO {
     private Long orderId;
     private List<OrderItemResponseDTO> orderItems = new ArrayList<>();
     private LocalDateTime orderDate;
-    private Double sellerAmount;
+    private BigDecimal sellerAmount;
     private OrderStatus orderStatus;
 }
