+++ src/main/java/com/ecommerce/project/repositories/OrderRepository.java
@@ -9,13 +9,14 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 @Repository
 public interface OrderRepository extends JpaRepository<Order, Long> {
 
     @Query("SELECT COALESCE(SUM(o.totalAmount), 0) FROM Order o")
-    Double getTotalRevenue();
+    BigDecimal getTotalRevenue();
 
     @EntityGraph(attributePaths = {
             "orderItems", "orderItems.product", "payment"
